--- liferay-workspace/wars/jsp-portlet-war/src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,11 +1,0 @@
-author=Liferay, Inc.
-change-log=
-licenses=Apache
-liferay-versions=7.0.0+
-long-description=
-module-group-id=liferay
-module-incremental-version=1
-name=Liferay MVC
-page-url=http://www.liferay.com
-short-description=
-tags=
