--- liferay-workspace/wars/jsp-portlet-war/src/main/webapp/WEB-INF/liferay-portlet.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE liferay-portlet-app PUBLIC "-//Liferay//DTD Portlet Application 7.0.0//EN" "http://www.liferay.com/dtd/liferay-portlet-app_7_0_0.dtd">
-
-<liferay-portlet-app>
-	<portlet>
-		<portlet-name>jsp-portlet-war</portlet-name>
-		<icon>/icon.png</icon>
-		<instanceable>true</instanceable>
-		<requires-namespaced-parameters>false</requires-namespaced-parameters>
-		<header-portlet-css>/css/main.css</header-portlet-css>
-		<css-class-wrapper>portlet-jsp</css-class-wrapper>
-	</portlet>
-	<role-mapper>
-		<role-name>administrator</role-name>
-		<role-link>Administrator</role-link>
-	</role-mapper>
-	<role-mapper>
-		<role-name>guest</role-name>
-		<role-link>Guest</role-link>
-	</role-mapper>
-	<role-mapper>
-		<role-name>power-user</role-name>
-		<role-link>Power User</role-link>
-	</role-mapper>
-	<role-mapper>
-		<role-name>user</role-name>
-		<role-link>User</role-link>
-	</role-mapper>
-</liferay-portlet-app>
