+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,0 +1,11 @@
+author=Liferay, Inc.
+change-log=
+licenses=Apache
+liferay-versions=7.0.0+
+long-description=
+module-group-id=liferay
+module-incremental-version=1
+name=JSP WAR Portlet
+page-url=http://www.liferay.com
+short-description=
+tags=
