+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/WEB-INF/portlet.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0"?>
+
+<portlet-app xmlns="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" version="2.0">
+	<portlet>
+		<portlet-name>jsp-war-portlet</portlet-name>
+		<display-name>JSP WAR Portlet</display-name>
+		<portlet-class>com.liferay.blade.samples.portlet.JSPWARPortlet</portlet-class>
+		<init-param>
+			<name>template-path</name>
+			<value>/</value>
+		</init-param>
+		<init-param>
+			<name>view-template</name>
+			<value>/view.jsp</value>
+		</init-param>
+		<expiration-cache>0</expiration-cache>
+		<supports>
+			<mime-type>text/html</mime-type>
+		</supports>
+		<resource-bundle>content.Language</resource-bundle>
+		<portlet-info>
+			<title>Liferay MVC</title>
+			<short-title>Liferay MVC</short-title>
+			<keywords>Liferay MVC</keywords>
+		</portlet-info>
+		<security-role-ref>
+			<role-name>administrator</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>guest</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>power-user</role-name>
+		</security-role-ref>
+		<security-role-ref>
+			<role-name>user</role-name>
+		</security-role-ref>
+	</portlet>
+</portlet-app>
