+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/WEB-INF/tld/liferay-security.tld
