+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/WEB-INF/web.xml
