+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/edit_foo.jsp
