+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/foo_action.jsp
