+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/icon.png
