+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/init.jsp
