+++ liferay-workspace/wars/jsp-war-portlet/src/main/webapp/view.jsp
