+++ maven/apps/jsp-war-portlet/sourcechecks-suppressions.xml
