+++ maven/apps/jsp-war-portlet/src/main/webapp/WEB-INF/liferay-display.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<!DOCTYPE display PUBLIC "-//Liferay//DTD Display 7.0.0//EN" "http://www.liferay.com/dtd/liferay-display_7_0_0.dtd">
+
+<display>
+	<category name="category.sample">
+		<portlet id="jsp-war-portlet" />
+	</category>
+</display>
