+++ maven/apps/jsp-war-portlet/src/main/webapp/icon.png
