+++ maven/apps/jsp-war-portlet/src/main/webapp/init.jsp
