+++ maven/apps/jsp-war-portlet/src/main/webapp/view.jsp
