+++ maven/apps/pom.xml
@@ -23,7 +23,7 @@
 		<module>freemarker-portlet</module>
 		<module>greedy-policy-option-portlet</module>
 		<module>jsp-portlet</module>
-		<module>jsp-portlet-war</module>
+		<module>jsp-war-portlet</module>
 		<module>osgi-portlet</module>
 		<module>render-command-portlet</module>
 		<module>resource-command-portlet</module>
